api/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
dj-database-url>=2.2,<3.0
django-cors-headers>=4.4,<5.0
gunicorn>=22.0,<23.0
psycopg2-binary>=2.9,<3.0
neomodel>=6.0,<7.0
neo4j>=5.0,<7.0
anthropic>=0.28,<1.0
openai>=1.0,<2.0
jsonschema>=4.20,<5.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
